startup-intelligence-agent/backend/requirements.txt
# FastAPI and web server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.23

# Data validation
pydantic>=2.5.0

# HTTP client
httpx>=0.25.1

# Environment variables
python-dotenv>=1.0.0

# LLM integrations
openai>=1.3.0
anthropic>=0.7.0

# Logging and utilities
python-dateutil>=2.8.2

# E2B Sandbox integration
e2b>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
httpx>=0.25.1
requests>=2.32.0

